External Libraries: Update the Requests library to version 1.8.1.
This is a minor update, with the most important change being that the version constant in the `Requests` class is now correct.

Release notes: https://github.com/WordPress/Requests/releases/tag/v1.8.1

For a full list of changes in this update, see the Requests GitHub:
https://github.com/WordPress/Requests/compare/v1.8.0...v1.8.1

Follow-up to [50842].

Props kapilpaul, jrf, schlessera, mbabker.
Fixes #53334.

git-svn-id: https://develop.svn.wordpress.org/trunk@51078 602fd350-edb4-49c9-b593-d223f7449a82